requirements.txt
requests
python-whois
tqdm